requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
requests>=2.28.2
httpx>=0.24.0
beautifulsoup4>=4.12.0
selenium>=4.9.0
webdriver-manager>=3.8.6
openai>=1.0.0
langchain>=0.0.200
python-dotenv>=1.0.0 